attendance system is updated

+++ src/main/java/np/edu/nast/payroll/Payroll/controller/AttendanceController.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/attendance")
-@CrossOrigin(origins = "http://localhost:5173") // Matches your frontend port
+@CrossOrigin(origins = "http://localhost:5173")
 public class AttendanceController {
 
     private final AttendanceService attendanceService;
@@ -22,14 +22,12 @@ public class AttendanceController {
     }
 
     @PutMapping("/{id}")
-    public Attendance update(@PathVariable Long id, @RequestBody Attendance attendance) {
-        // Ensuring ID conversion to match your service's expected type
-        return attendanceService.updateAttendance(id.intValue(), attendance);
+    public Attendance update(@PathVariable Integer id, @RequestBody Attendance attendance) {
+        return attendanceService.updateAttendance(id, attendance);
     }
 
     @GetMapping("/employee/{empId}")
     public List<Attendance> getByEmployee(@PathVariable Integer empId) {
         return attendanceService.getAttendanceByEmployee(empId);
     }
-
 }
